refactor: move functions to helpers.tsx
