+++ packages/ai-lab/src/components/AILabImage/AILabImage.tsx
@@ -3,7 +3,12 @@ import * as tf from '@tensorflow/tfjs';
 import { Performance, PerformanceInfo, perfInfo } from '../../performance';
 import { ImageProps, ModelConfig, Results } from '../../types';
 import { AILabObjectDetectionUI } from '../AILabObjectDetectionUI';
-import { CLASSES } from '../labels';
+import {
+  getInferenceData,
+  getModelDetections,
+  predictClassification,
+  predictSSD,
+} from '../helpers';
 
 const defaultModelConfig: ModelConfig = {
   modelType: 'ssd',
@@ -13,110 +18,6 @@ const defaultModelConfig: ModelConfig = {
   nmsActive: true,
   topK: 5,
 };
-
-async function ssdModelDetection(results: Results, config: ModelConfig) {
-  // Get a clean tensor of top indices
-  const prominentDetection = tf.topk((results as tf.Tensor<tf.Rank>[])[0]);
-
-  const justBoxes = (results as tf.Tensor<tf.Rank>[])[1].squeeze<
-    tf.Tensor<tf.Rank.R2>
-  >();
-
-  const justValues = prominentDetection.values.squeeze<tf.Tensor<tf.Rank.R1>>();
-
-  const { threshold, maxResults = 20, iouThreshold, nmsActive } = config;
-
-  // Move results back to JavaScript in parallel
-  const [maxIndices, scores, boxes] = await Promise.all([
-    prominentDetection.indices.data(),
-    justValues.array(),
-    justBoxes.array(),
-  ]);
-
-  // https://arxiv.org/pdf/1704.04503.pdf, use Async to keep visuals
-  const nmsDetections = await tf.image.nonMaxSuppressionWithScoreAsync(
-    justBoxes,
-    justValues,
-    maxResults, //maxBoxes
-    iouThreshold,
-    threshold,
-    nmsActive ? 1 : 0 // 0 is normal NMS, 1 is Soft-NMS for overlapping support
-  );
-
-  const detections = await nmsDetections.selectedIndices.data();
-
-  tf.dispose([nmsDetections.selectedIndices, nmsDetections.selectedScores]);
-
-  return { detections, maxIndices, scores, boxes };
-}
-
-async function classificationModelDetection(
-  results: Results,
-  config: ModelConfig
-) {
-  const values = await (results as tf.Tensor2D).data();
-  const valuesAndIndices = [];
-  for (let i = 0; i < values.length; i++) {
-    valuesAndIndices.push({ value: values[i], index: i });
-  }
-
-  valuesAndIndices.sort((a, b) => {
-    return b.value - a.value;
-  });
-
-  const topkValues = new Float32Array(config.topK!);
-  const topkIndices = new Int32Array(config.topK!);
-  for (let i = 0; i < config.topK!; i++) {
-    topkValues[i] = valuesAndIndices[i].value;
-    topkIndices[i] = valuesAndIndices[i].index;
-  }
-
-  const { threshold, maxResults = config.topK! } = config;
-  const topClassesAndProbs = [];
-  for (let i = 0; i < maxResults; i++) {
-    topClassesAndProbs.push(topkValues[i]);
-  }
-
-  const finalResults = topClassesAndProbs.filter(function (r) {
-    return r > threshold!;
-  });
-
-  return finalResults;
-}
-
-async function predictSSD(
-  tensor: tf.Tensor3D,
-  model: tf.GraphModel | tf.LayersModel
-) {
-  // SSD Mobilenet single batch
-
-  const readyfied = tf.expandDims(tensor, 0);
-  const res = await (model as tf.GraphModel).executeAsync(readyfied);
-
-  return res;
-}
-
-async function predictClassification(
-  tensor: tf.Tensor3D,
-  model: tf.GraphModel | tf.LayersModel,
-  size: number
-) {
-  const readyfied = tensor.toFloat().div(255);
-  let resized = readyfied;
-  if (tensor.shape[0] !== size || tensor.shape[1] !== size) {
-    const alignCorners = true;
-    resized = tf.image.resizeBilinear(
-      readyfied as tf.Tensor3D,
-      [size, size],
-      alignCorners
-    );
-  }
-  // Reshape to a single-element batch so we can pass it to predict.
-  const batched = resized.reshape([1, size, size, 3]);
-  const res = model.predict(batched);
-
-  return res;
-}
 
 export const AILabImage = ({
   model,
@@ -159,34 +60,6 @@ export const AILabImage = ({
     }
   };
 
-  async function getInferData(res: Results) {
-    // Store Box Detections
-    if (modelConfig?.modelType === 'ssd') {
-      const { detections, maxIndices, scores, boxes } = await ssdModelDetection(
-        res,
-        {
-          ...defaultModelConfig,
-          ...modelConfig,
-        }
-      );
-      setDetectionResults({ detections, maxIndices, scores, boxes });
-
-      const ssdInferData = Array.from(detections).map((d) => ({
-        detectedClass: CLASSES[maxIndices[d]],
-        detectedScore: scores[d],
-      }));
-
-      onInference?.(ssdInferData);
-    } else {
-      const finalResults = await classificationModelDetection(res, {
-        ...defaultModelConfig,
-        ...modelConfig,
-      });
-
-      onInference?.(finalResults);
-    }
-  }
-
   useEffect(() => {
     tf.ready().then(() => {
       setIsTFReady(true);
@@ -225,9 +98,14 @@ export const AILabImage = ({
   }, [src, isTFReady]);
 
   useEffect(() => {
-    if (results) {
-      getInferData(results);
-    }
+    (async function () {
+      if (results) {
+        const detections = await getModelDetections(results, modelConfig);
+        const inferences = await getInferenceData(detections);
+        setDetectionResults(detections);
+        onInference?.(inferences);
+      }
+    })();
   }, [modelConfig, results]);
 
   return (
